pgm2 - reshuffle things a bit (split into driver/video etc.), make a few notes, correct main ram size
